Made UI responsive, and added drop down menu for mobile that's impossible to toggle because ruby on rails is kinda.... ugh

+++ db/migrate/20180124232710_create_store_transactions.rb
@@ -2,7 +2,7 @@ class CreateStoreTransactions < ActiveRecord::Migration[5.1]
   def change
     create_table :store_transactions do |t|
       t.datetime :dateTime
-      t.integer :charge
+      t.float :charge
       t.belongs_to :customerID, foreign_key: true
       t.belongs_to :productID, foreign_key: true
